move resources page to dynamic route.
