+++ src/app/(main)/my-voyage/[teamId]/voyage-resources/components/EmptyBanner.tsx
