+++ src/app/(main)/my-voyage/[teamId]/voyage-resources/components/ModalSection.tsx
