+++ src/app/(main)/my-voyage/[teamId]/voyage-resources/components/ResourceInput.tsx
