+++ src/app/(main)/my-voyage/[teamId]/voyage-resources/components/fixtures/resources.ts
