+++ src/components/providers/ModalProvider.tsx
@@ -7,6 +7,8 @@ import Example1Modal from "@/components/modals/Example1Modal";
 import Example2Modal from "@/components/modals/Example2Modal";
 import FeatureModal from "@/components/modals/FeatureModal";
 import ErrorModal from "@/components/modals/ErrorModal";
+import DeleteModal from "@/components/modals/DeleteModal";
+import ViewModal from "@/components/modals/ViewModal";
 
 export default function ModalProvider() {
   const [isMounted, setIsMounted] = useState(false);
@@ -24,6 +26,8 @@ export default function ModalProvider() {
       <FeatureModal />
       <ErrorModal />
       <GettingHelpModal />
+      <DeleteModal />
+      <ViewModal />
     </>
   );
 }
